# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/tracks": {
      "get": {
        "operationId": "AdminTracksController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "HIDDEN",
                "PUBLISHED"
              ]
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artistId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTracksListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tracks (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      },
      "post": {
        "operationId": "AdminTracksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateTrackRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create track metadata (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}": {
      "get": {
        "operationId": "AdminTracksController_detail",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get track detail (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      },
      "patch": {
        "operationId": "AdminTracksController_update",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateTrackRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update track metadata (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}/original-upload-url": {
      "post": {
        "operationId": "AdminTracksController_originalUploadUrl",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackUploadUrlResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed upload URL for original audio (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}/preview-upload-url": {
      "post": {
        "operationId": "AdminTracksController_previewUploadUrl",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackUploadUrlResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed upload URL for preview audio (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}/preview-playback-url": {
      "get": {
        "operationId": "AdminTracksController_previewPlaybackUrl",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackPlaybackUrlResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed playback URL for preview audio (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}/publish": {
      "patch": {
        "operationId": "AdminTracksController_publish",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish track (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
    "/admin/tracks/{trackId}/hide": {
      "patch": {
        "operationId": "AdminTracksController_hide",
        "parameters": [
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTrackResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hide track (admin)",
        "tags": [
          "Admin - Tracks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Musica API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TrackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artistId": {
            "type": "string"
          },
          "authorName": {
            "type": "object",
            "nullable": true
          },
          "genre": {
            "type": "object",
            "nullable": true
          },
          "duration": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "HIDDEN",
              "PUBLISHED"
            ]
          },
          "usageRights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "originalAudioKey": {
            "type": "object",
            "nullable": true
          },
          "previewAudioKey": {
            "type": "object",
            "nullable": true
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artistId",
          "status",
          "usageRights",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminTracksListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "PaginationInfoDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPrevPage": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasNextPage",
          "hasPrevPage"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfoDto"
          }
        },
        "required": [
          "pagination"
        ]
      },
      "AdminTracksListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminTracksListDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "requestId": {
            "type": "string",
            "example": "2fefcbd8-0a70-4c9d-8e86-e88f7b0f5c5a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-20T00:00:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "meta",
          "requestId",
          "timestamp"
        ]
      },
      "AdminCreateTrackRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "artistId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "usageRights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "artistId"
        ]
      },
      "AdminTrackResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/TrackDto"
          },
          "requestId": {
            "type": "string",
            "example": "2fefcbd8-0a70-4c9d-8e86-e88f7b0f5c5a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-20T00:00:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "requestId",
          "timestamp"
        ]
      },
      "AdminUpdateTrackRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "artistId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "usageRights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminTrackUploadUrlResponseDataDto": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "fileKey": {
            "type": "string"
          }
        },
        "required": [
          "uploadUrl",
          "fileKey"
        ]
      },
      "AdminTrackUploadUrlResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminTrackUploadUrlResponseDataDto"
          },
          "requestId": {
            "type": "string",
            "example": "2fefcbd8-0a70-4c9d-8e86-e88f7b0f5c5a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-20T00:00:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "requestId",
          "timestamp"
        ]
      },
      "AdminTrackPlaybackUrlResponseDataDto": {
        "type": "object",
        "properties": {
          "playbackUrl": {
            "type": "string"
          }
        },
        "required": [
          "playbackUrl"
        ]
      },
      "AdminTrackPlaybackUrlResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminTrackPlaybackUrlResponseDataDto"
          },
          "requestId": {
            "type": "string",
            "example": "2fefcbd8-0a70-4c9d-8e86-e88f7b0f5c5a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-20T00:00:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "requestId",
          "timestamp"
        ]
      },
...
    }
  }
}